+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unrealese]
 
+### Added
+- added the ability to update tokens when executing a go event
+
 ### Fixed
 - Multiline. Moved the barplots to the background